+++ assets/javascripts/ListFilter.js
@@ -1,19 +1,14 @@
 import { convertNodeListToArray } from './utils.js';
 
 export class ListFilter{
-    static makeFromForm($form){
-      if($form){
-        const lf = new ListFilter($form);
-        lf.init();
-        return lf;
-      }
-      throw new Error('ListFilter.makeFilterable requires a form element');
-    }
-
     constructor($form) {
+        if(!$form){
+          throw new Error('ListFilter requires a form element');
+        }
         this.$form = $form;
         this.filterTimeout = null;
         this.$noMatches = document.querySelector('.dl-list-filter__no-filter-match');
+        this.init();
     }
 
     init(params){
@@ -26,7 +21,7 @@ export class ListFilter{
 
         const $input = $form.querySelector('input');
         const boundFilterViaTimeout = this.filterViaTimeout.bind(this);
-        $input.addEventListener('keyup', boundFilterViaTimeout);
+        $input.addEventListener('input', boundFilterViaTimeout);
 
         // make sure no matches message is initially hidden
         this.$noMatches.classList.add('js-hidden');
@@ -35,13 +30,13 @@ export class ListFilter{
     filterViaTimeout(e){
         clearTimeout(this.filterTimeout);
 
-        const boundListFilter = this.ListFilter.bind(this);
+        const boundListFilter = this.filterListItems.bind(this);
         this.filterTimeout = setTimeout(function () {
           boundListFilter(e);
         }, 200);
     }
 
-    ListFilter(e){
+    filterListItems(e){
         const itemsToFilter = convertNodeListToArray(document.querySelectorAll('[data-filter="item"]'));
         const listsToFilter = convertNodeListToArray(document.querySelectorAll('[data-filter="list"]'));
         const searchTerm = e.target.value;
@@ -49,7 +44,7 @@ export class ListFilter{
         const boundMatchSearchTerm = this.matchSearchTerm.bind(this);
         itemsToFilter
           .filter(function ($item) {
-            return boundMatchSearchTerm($item, searchTerm)
+            return !boundMatchSearchTerm($item, searchTerm)
           })
           .forEach(function (item) {
             item.classList.add('js-hidden');
@@ -76,9 +71,9 @@ export class ListFilter{
         item.classList.remove('js-hidden');
         var searchTermRegexp = new RegExp(term.replace(/[.*+?^${}()|[\]\\]/g, '\\$&'), 'i');
         if (searchTermRegexp.exec(contentToMatchOn) !== null) {
-          return false
+          return true
         }
-        return true
+        return false
     }
 
     updateListCounts(lists){
@@ -103,9 +98,6 @@ export class ListFilter{
           }
 
           totalMatches += matchingCount;
-
-          var filteredEvent = new CustomEvent('list:filtered');
-          list.dispatchEvent(filteredEvent);
         });
 
         // if no results show message