+++ tests/utils/mockUtils.js
@@ -56,6 +56,9 @@ const mapMock = {
         }
     }),
     moveLayer: vi.fn(),
+    getCanvas: vi.fn().mockImplementation(() => {
+        return domElementMock;
+    })
 };
 
 const mapControllerMock = {
@@ -87,7 +90,7 @@ const domElementMock = {
     addEventListener: vi.fn(),
     removeEventListener: vi.fn(),
     setAttribute: vi.fn(),
-    focus: vi.fn(),
+    focus: vi.fn()
 }
 
 const popupMock = {
@@ -137,13 +140,18 @@ export const stubGlobalWindow = (pathname = 'http://localhost', hash = '') => {
         },
         history: {
             pushState: vi.fn(),
+            replaceState: vi.fn(),
         },
+        setTimeout: vi.fn().mockImplementation((callback) => {
+            callback();
+        }),
+        clearTimeout: vi.fn(),
     })
 }
 
 let urlParams = [];
 
-export const stubGlobalUrl = (urlParams = [], urlParamSize = 0) => {
+export const stubGlobalUrl = (urlParams = []) => {
     urlParams = urlParams || [];
     const deleteMock = vi.fn().mockImplementation((key) => {
         urlParams = urlParams.filter((param) => {
@@ -177,7 +185,7 @@ export const stubGlobalUrl = (urlParams = [], urlParamSize = 0) => {
                     })
                     return toReturn;
                 }),
-                size: urlParamSize,
+                size: urlParams.length,
             }
         }
     }))
@@ -192,10 +200,16 @@ export const stubGlobalDocument = (location = 'http://localhost:3000/?layers=lay
         querySelector: vi.fn().mockImplementation(() => {
             return domElementMock
         }),
+        querySelectorAll: vi.fn().mockImplementation(() => {
+            return [domElementMock, domElementMock]
+        }),
         location: location,
         getElementById: vi.fn().mockImplementation(() => {
             return domElementMock
         }),
+        body: {
+            removeChild: vi.fn(),
+        }
     })
 }