+++ changeLog.md
@@ -7,12 +7,107 @@
     - ...
     ### Why was this change made?
     - [explanation]
-    ***
+
     <br />
 -->
 
 # ChangeLog
 <br>
+
+## 19-10-2023
+### What's new
+- Added unit tests for the list filter
+- Added acceptance tests for the list filter
+- Fixed panning issue on the map page by changing the handlemove to only be called on moveend instead of move
+### Why was this change made?
+- to ensure this code is robust
+- because this user journey was documented as part of a recent group session we did
+- because this was causing the map to crash on some browsers
+
+## 17-10-2023
+### What's new
+- Added acceptance test for using all the filters on the search page
+### Why was this change made?
+- because this user journey was documented as part of a recent group session we did
+
+## 12-10-2023
+### What's new
+- Added acceptance test for navigating to the organisations page from the entity page
+### Why was this change made?
+- because this user journey was documented as part of a recent group session we did
+
+## 11-10-2023
+### What's new
+- Updated the map testing code to use the new map page object model
+- Added commented test for the map page, to be fixed at a later date
+### Why was this change made?
+- because this user journey was documented as part of a recent group session we did
+### Additional changes
+- Added an additional entity to the test data, and fixed tests to account for this
+- Fixed some accessibility issues with the entity page
+- Period selected value not based of how many checkboxes checked instead of hardcoded to be 1
+- Make sure datasets are in alphabetical order on the guidance pages
+
+## 10-10-2023
+### What's new
+- Optimized the queries in get_entity_count and get_entity_search
+### Why was this change made?
+- The production database was using too much local storage performing some of the most common queries.
+
+## 10-10-2023
+### What's new
+- Added an acceptance test that tests navigation to a dataset page
+- Added an acceptance test that tests loading the swagger editor with the correct openAPI json
+### Why was this change made?
+- because these user journeys were documented as part of a recent group session we did
+
+## 09-10-2023
+### What's new
+- Added tests to ensure each and every page request returns a successful response
+- Refactored the tests to remove duplicate code
+### Why was this change made?
+- To ensure we don't introduce any broken pages to the live site
+
+## 06-10-2023
+### What's new
+- entities filter renamed to period
+- period filter now uses checkboxes
+- period filter clear now works as expected
+- period value (historical/current) now indicated on the entity
+### Why was this change made?
+- To ensure user don't mistake historical entities for current entities
+
+## 03-10-2023
+### What's new
+- Fixed the layer controls component for safari, as the checkboxes were not changing state
+### Why was this change made?
+- So the layer controls work on safari
+
+## 29-09-2023
+### What's new
+- Updated the rendering of a polygon layer to also check for point data and render a circle where appropriate
+### Why was this change made?
+- to handle the new NSIP dataset
+
+## 29-09-2023
+### What's new
+- Setup axe playwright accessibility testing for all pages
+- Fixed any issues found by the axe tests
+- updated git actions to run the axe tests
+### Why was this change made?
+- To ensure that future changes don't break accessibility
+
+## 29-09-2023
+### What's new
+- decrease wales obscure opacity
+### Why was this change made?
+- Some data we have is on wales, so we need to obscure it less
+
+## 26-09-2023
+### What's new
+- Updated the base-tile set style changing min and max zoom levels for layers
+### Why was this change made?
+- To avoid the map being to cluttered and to make it easier to see the data
 
 ## 22-09-2023
 ### What's new
@@ -53,7 +148,6 @@
 - The cursor pointer makes it more obvious that the feature is clickable
 - To avoid people thinking that our data is owned by OS
 - To make it easier to share a specific location on the map
-***
 <br />
 
 ## 13-09-2023
@@ -63,7 +157,6 @@
 - the back button now correctly maintains the previous page state on the edge browser
 ### Why was this change made?
 - Samantha requested that the back button should simulate the window history back
-***
 <br />
 
 ## 13-09-2023
@@ -72,13 +165,11 @@
 ### Why was this change made?
 - The previous layer controls wouldn't correctly appear in when the map was full screened
 - this resolves tech debt surrounding this component
-***
 <br />
 
 ## 11-09-2023
 ### What's new
 - fix the location search bar on the search page
-***
 <br />
 
 ## 07-09-2023
@@ -86,7 +177,6 @@
 - map popups now dont have a scroll event listener
 ### Why was this change made?
 - previously, the scroll event listener would cause the page to scroll when zooming on the popup
-***
 <br />
 
 ## 07-09-2023
@@ -94,7 +184,6 @@
 - phase banner now has appropriate margin on the map page.
 ### Why was this change made?
 - previously the phase banner looked out of place on the map page
-***
 <br />
 
 ## 07-09-2023
@@ -104,7 +193,6 @@
 ### Why was this change made?
 - showing historical entities from the start is confusing
 - unnecessary to show typologies that have no entities
-***
 <br />
 
 ## 06-09-2023
@@ -112,7 +200,6 @@
 - hide cookie banner if js is disabled
 ### Why was this change made?
 - we can't store cookies without js, so no need to show this
-***
 <br />
 
 ## 06-09-2023
@@ -124,7 +211,6 @@
 - maptiler would eventually revoke our key again, os maps is free for us to use
 - we want to try and keep our os map key and secret secure so oauth2 helps us protect this
 - 3D polygons look better than 2D polygons
-***
 <br />
 
 
@@ -133,7 +219,6 @@
 - Fix the search bar on the datasets filters
 ### Why was this change made?
 - To make the search bar work as intended
-***
 <br />
 
 ## 31-08-2023
@@ -143,7 +228,6 @@
 - update about page to use more syntactical html
 ### Why was this change made?
 - To make the page more accessible
-***
 <br />
 
 ## 30-08-2023
@@ -152,7 +236,6 @@
 - Update the service status page to remove separators from the accessibility dom
 ### Why was this change made?
 - To make the pages more accessible
-***
 <br />
 
 ## 25-08-2023
@@ -160,7 +243,6 @@
 - Make banner research sign up link open in a new tab
 ### Why was this change made?
 - Request made by MJ
-***
 <br />
 
 ## 25-08-2023
@@ -169,7 +251,6 @@
 ### Why was this change made?
 - The pages don't always follow a linier structure stable for breadcrumbs
 - A back button is more consistent with the site design
-***
 <br />
 
 ## 25-08-2023
@@ -178,7 +259,6 @@
 ### Why was this change made?
 - Enables better map configuration and easier reuse
 - Resolves some tech debt
-***
 <br />
 
 ## 24-08-2023
@@ -186,7 +266,6 @@
 - Move nav bar bellow page title on map page
 ### Why was this change made?
 - To maintain consistency across the site
-***
 <br />
 
 ## 24-08-2023
@@ -195,7 +274,6 @@
 ### Why was this change made?
 - To better make the banner stand out
 - Also to link users to sign up for user research
-***
 <br />
 
 ## 24-08-2023
@@ -204,7 +282,6 @@
 ### Why was this change made?
 - To remove the need to hard code the measurment ID in the javascript code
 - To enable the use of Google Analytics in the development environment
-***
 <br />
 
 ## 16-08-2023
@@ -212,12 +289,10 @@
 - changed the base tile set and maptiler access key to George Goodall's personal key
 ### Why was this change made?
 - Paul Smith's key was revoked
-***
 
 ## 02-08-2023
 ### What's new
 - changed guidance string to a header on the article-4-direction specification
-***
 <br />
 
 ## 12-07-2023
@@ -226,4 +301,5 @@
 - Added a template for the change log
 ### Why was this change made?
 - To keep track of changes made to the project
-***
+
+<br />